HT-2907: Move HTItemClusterGetter to Clustering module
