+++ lib/clustering/ht_item_cluster_getter.rb
